+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+Changes:
+*   Add example running attacks with only csv probabilities supplied ([#353](https://github.com/AI-SDC/SACRO-ML/pull/353))
+
 ## Version 1.4.0 (Jul 4, 2025)
 
 Changes: